app restructure
- restructure  and reorganize  all folders
- rename folders and files to a naming styles
- update all yarn packages to latest
- remove all unnecessary file
- switch from css to scss

[WIP]
