+++ src/App/__test__/app.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import App from './App'
+import App from '..'
 
 test('renders learn react link', () => {
   render(<App />)
