+++ src/Assets/index.scss
@@ -1,0 +1,7 @@
+body {
+  margin: 0;
+  padding: 0;
+}
+@import './styles/index.scss';
+@import 'App/app.scss';
+@import 'Components/';
