+++ src/Pages/BuyFlow/index.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react'
-import AgeStep from './AgeStep'
-import EmailStep from './EmailStep'
-import SummaryStep from './SummaryStep'
+import AgeStep from './ageStep'
+import EmailStep from './emailStep'
+import SummaryStep from './summaryStep'
 
-interface BuyflowProps {
+interface BuyFlowProps {
   productId: ProductIds
 }
 
@@ -15,7 +15,7 @@ const PRODUCT_IDS_TO_NAMES = {
   [ProductIds.devIns]: 'Developer Insurance',
 }
 
-const Buyflow: React.FC<BuyflowProps> = (props) => {
+const BuyFlow: React.FC<BuyFlowProps> = (props) => {
   const [currentStep, setStep] = useState('email')
   const [collectedData, updateData] = useState({
     email: '',
@@ -39,4 +39,4 @@ const Buyflow: React.FC<BuyflowProps> = (props) => {
   )
 }
 
-export default Buyflow
+export default BuyFlow
