+++ src/Pages/index.tsx
@@ -1,0 +1,3 @@
+import BuyFlow from './BuyFlow'
+
+export { BuyFlow }
